[NEW #241] Added Support for MySQL

+++ yafsrc/YAF.Web/Controls/OnlineStatusIcon.cs
@@ -3,7 +3,7 @@
  * Copyright (C) 2006-2013 Jaben Cargman
  * Copyright (C) 2014-2021 Ingo Herbote
  * https://www.yetanotherforum.net/
- * 
+ *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -101,7 +101,7 @@ namespace YAF.Web.Controls
             }
             else
             {
-                if (activeUsers.Any(x => (int)x.UserID == this.UserId && !x.IsActiveExcluded))
+                if (activeUsers.Any(x => x.UserID == this.UserId && !x.IsActiveExcluded))
                 {
                     // online
                     writer.WriteAttribute(HtmlTextWriterAttribute.Class.ToString(), "align-middle text-success user-online me-1");
